+++ CHANGELOG.md
@@ -3,6 +3,6 @@
     [logstash-filter-jdbc_static](https://github.com/logstash-plugins/logstash-filter-jdbc_static)
   - For Changelog of individual plugins, see:
     - [JBDC Input version 4.3.19](https://github.com/logstash-plugins/logstash-input-jdbc/blob/v4.3.19/CHANGELOG.md)
-    - [JDBC Static filter version 1.0.7](https://github.com/logstash-plugins/logstash-filter-jdbc_static/blob/v1.0.7/CHANGELOG.md)
+    - [JDBC Static filter version 1.1.0](https://github.com/logstash-plugins/logstash-filter-jdbc_static/blob/v1.1.0/CHANGELOG.md)
     - [JDBC Streaming filter version 1.0.10](https://github.com/logstash-plugins/logstash-filter-jdbc_streaming/blob/v1.0.10/CHANGELOG.md)